remove identity processor
