--- data_processor/steps/coarse_processing/data_process/process_identity.py
@@ -1,34 +1,0 @@
-# !/usr/bin/env python3
-
-# Copyright (c) 2025 PaddlePaddle Authors. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-"""
-Identity processor
-"""
-
-
-class IdentityProcessor:
-    """
-    Identity processor
-    """
-
-    def __init__(self, *args, **kwargs):
-        return
-
-    def process(self, schema, **kwargs):
-        """
-        identity processor
-        """
-        return schema
